Apply code formatting suggestions from review
Should help in regards to readability, as lambdaurora suggested these changes would. Removal of the second `if (!fsHelper)` check was skipped, as while I suspect there's a more elegant way to write that block of code, doing that would actually break the current behavior—that second if statement is necessary for the current loose files warning to work correctly, and not be displayed on any loose file that _doesn't_ start with `._`.

Co-authored-by: LambdAurora <aurora42lambda@gmail.com>